Updated variables to use more appropriate types
Renamed functions for better readability

Edited on_tick() return to Running - forcing overrides to set return manually

Added a merge function to gworld_model

Updated is_goal_reached logic and switch from for loop to ranges::any_of
Fixed a bug for non-equal preconditions

Fixed bug where calls to plan() would always use the search algorithm even if the start state satisfied the goal state

+++ gplan_result.h
@@ -23,7 +23,7 @@ struct gplan_result
     std::vector<gaction::const_ptr> actions;
     gplan_status status = gplan_status::Success;
 
-    int nodes_expanded = 0;
+    int64_t  nodes_expanded = 0;
     int final_cost = 0;
     int planning_time_ms = 0;
